restructure component logic
