+++ app/(auth)/layout.tsx
@@ -4,10 +4,8 @@ export default function RootLayout({
     children: React.ReactNode;
 }>) {
     return (
-        <html lang="en">
-            <body>
-                {children}
-            </body>
-        </html>
+        <div>
+            {children}
+        </div>
     );
 }
