+++ components/dashboard/Dashboard.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from "react";
 import { User } from "@supabase/supabase-js";
 import { createClientComponentClient } from "@supabase/auth-helpers-nextjs";
 
-export function AdminDashboard() {
+export function Dashboard() {
   const [user, setUser] = useState<User | null>(null);
   const [error, setError] = useState<string | null>(null);
   const supabase = createClientComponentClient();
@@ -13,32 +13,11 @@ export function AdminDashboard() {
         const {
           data: { user: authUser },
         } = await supabase.auth.getUser();
-        if (!authUser) {
-          throw new Error("User not found");
-        }
 
-        const { data: userData, error: userError } = await supabase
-          .from("users")
-          .select("*")
-          .eq("id", authUser.id)
-          .single();
-
-        if (userError) {
-          throw new Error("Failed to fetch user data");
-        }
-
-        const combinedUser = {
-          ...authUser,
-          user_metadata: {
-            ...authUser.user_metadata,
-            role: userData.role,
-          },
-        };
-
-        setUser(combinedUser);
+        setUser(authUser)
 
       } catch (error) {
-        console.error("Error in AdminDashboard:", error);
+        console.error("Error in Dashboard:", error);
         setError(error instanceof Error ? error.message : "An error occurred");
       }
     };
@@ -59,7 +38,7 @@ export function AdminDashboard() {
   return (
     <div className="p-8">
       <div className="mb-4">
-        <h1 className="text-2xl font-bold">Admin Dashboard</h1>
+        <h1 className="text-2xl font-bold">Dashboard</h1>
       </div>
 
       <div className="p-6 bg-white rounded-lg shadow">
