+++ components/sidebar.tsx
@@ -3,8 +3,6 @@
 import Link from "next/link";
 import Image from "next/image";
 import { usePathname } from "next/navigation";
-import { useEffect, useState } from 'react';
-import { createClientComponentClient } from '@supabase/auth-helpers-nextjs';
 import type { Database } from '@/types/supabase';
 
 import { IconUser } from "./icon/IconUser";
@@ -12,67 +10,19 @@ import { IconSetting } from "./icon/IconSetting";
 import { IconNotification } from "./icon/IconNotification";
 import LogoutButton from "@/components/ui/logout-button";
 
-const defaultNavigation = [
-  { name: "Dashboard", href: "/dashboard", roles: ['admin', 'viewer'] },
-  { name: "Event Listing", href: "/events", roles: ['admin', 'collaborator'] },
-  { name: "User Management", href: "/users", roles: ['admin', 'viewer'] },
-];
+type NavigationItem = {
+  name: string;
+  href: string;
+  roles: (Database['public']['Tables']['users']['Row']['role'])[];
+};
 
-export function Sidebar() {
+type SidebarProps = {
+  navigationItems: NavigationItem[],
+  isLoading: boolean
+}
+
+export function Sidebar({ navigationItems, isLoading }: SidebarProps) {
   const pathname = usePathname();
-  const supabase = createClientComponentClient<Database>();
-
-  const [userId, setUserId] = useState<string | null>(null);
-  const [userRole, setUserRole] = useState<string | null>(null);
-  const [isLoading, setIsLoading] = useState(true);
-
-  useEffect(() => {
-    const fetchUser = async () => {
-      setIsLoading(true);
-      const { data: { user }, error: userError } = await supabase.auth.getUser();
-
-      if (user) {
-        setUserId(user.id);
-        setUserRole(user.user_metadata.role)
-      }
-
-      setIsLoading(false)
-    };
-
-    fetchUser();
-
-    const { data: { subscription } } = supabase.auth.onAuthStateChange((event) => {
-      console.log('Sidebar Auth event:', event);
-      fetchUser();
-    });
-
-    return () => {
-      subscription?.unsubscribe();
-    };
-  }, [supabase]);
-
-
-  type NavigationItem = {
-    name: string;
-    href: string;
-    roles: (Database['public']['Tables']['users']['Row']['role'])[];
-  };
-  let navigationItems: NavigationItem[] = [];
-
-  if (!isLoading && userRole) {
-    navigationItems = defaultNavigation
-      .filter(item => item.roles.includes(userRole))
-      .map(item => ({ ...item }));
-
-    if (userRole === 'collaborator' && userId) {
-      navigationItems.push({
-        name: "My Profile",
-        href: `/users/${userId}`,
-        roles: ['collaborator']
-      });
-    }
-  }
-
   return (
     <div className="fixed left-0 top-0 flex h-screen w-[260px] flex-col border-r border-gray-200 bg-white">
       <div className="p-6 mb-8 border-b border-gray-200">
